requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
distro==1.8.0
Django==4.1.4
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-api-key==2.2.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==3.2.0
kombu==5.2.4
MarkupSafe==2.1.1
packaging==22.0
paramiko==2.12.0
pdfkit==1.0.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
psycopg2-binary==2.8.5
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.19.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==5.4.1
redis==4.4.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
texttable==1.6.7
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.59.0
